Refactor run integrity scanner architecture
