+++ Sources/QuillCodeCore/RunIntegrityScanner.swift
@@ -1,91 +1,5 @@
 import Foundation
 
-/// The post-run integrity verdict for a completed agent run: was "finished" honest?
-///
-/// This mechanizes the never-skip-tests rule so a human (or the morning-triage inbox) does not have to
-/// read the transcript to trust the run. It is deliberately a THREE-state badge with a HIGH-PRECISION
-/// bias: `unverified` and `red` are only raised on positive evidence of a problem, never on the mere
-/// absence of tests. When in doubt the scanner prefers `verified` over a false alarm, because a badge
-/// that cries wolf is a badge nobody reads.
-public enum RunIntegrityVerdict: String, Codable, Sendable, Hashable {
-    /// Nothing dishonest was detected: any success claim is backed by a real successful test/command
-    /// result, and no test/verify command was left failing. This is also the verdict for a benign run
-    /// that made no success claims and left no failing tests standing.
-    case verified
-
-    /// A success claim ("tests pass", "all green", "verified") has NO successful test/command result
-    /// backing it, OR a test that was started/expected was silently skipped. An honest "we cannot vouch
-    /// for this" — not necessarily a failure, but not a checked green either.
-    case unverified
-
-    /// A test/verify command exited nonzero and the failure was left standing (never re-run green). The
-    /// run "finished" on top of a known-red check.
-    case red
-
-    /// Badge text surfaced in Activity and the finish notification.
-    public var badgeLabel: String {
-        switch self {
-        case .verified: return "VERIFIED"
-        case .unverified: return "UNVERIFIED"
-        case .red: return "RED"
-        }
-    }
-}
-
-/// A single reason the scanner reached its verdict, kept for inspectability (the badge tooltip, the
-/// notification body, and — crucially — debugging false positives). Each reason names the rule that
-/// fired so the rule set stays auditable.
-public struct RunIntegrityReason: Sendable, Hashable {
-    public enum Rule: String, Sendable, Hashable {
-        /// A test/verify command failed and was never re-run successfully. -> RED
-        case standingTestFailure
-        /// The model claimed success but no successful test/command result backs it. -> UNVERIFIED
-        case unbackedSuccessClaim
-        /// A test command was queued/started but produced no completed result (silently skipped). -> UNVERIFIED
-        case skippedTest
-        /// A success claim is backed by an actual successful test/command result. -> VERIFIED (evidence)
-        case backedSuccessClaim
-        /// A test/verify command failed but a later run of it (or a test command) passed. -> VERIFIED (evidence)
-        case failureThenRerunGreen
-    }
-
-    public var rule: Rule
-    public var detail: String
-
-    public init(rule: Rule, detail: String) {
-        self.rule = rule
-        self.detail = detail
-    }
-}
-
-/// The full result of a run-integrity scan: the badge verdict plus the reasons behind it.
-public struct RunIntegrityReport: Sendable, Hashable {
-    public var verdict: RunIntegrityVerdict
-    public var reasons: [RunIntegrityReason]
-
-    public init(verdict: RunIntegrityVerdict, reasons: [RunIntegrityReason] = []) {
-        self.verdict = verdict
-        self.reasons = reasons
-    }
-
-    /// A one-line, human-readable summary for the notification body / badge tooltip. Never empty.
-    public var summaryLine: String {
-        switch verdict {
-        case .verified:
-            if let backed = reasons.first(where: { $0.rule == .backedSuccessClaim || $0.rule == .failureThenRerunGreen }) {
-                return backed.detail
-            }
-            return "No standing test failures or unbacked claims."
-        case .unverified:
-            return reasons.first(where: { $0.rule == .unbackedSuccessClaim || $0.rule == .skippedTest })?.detail
-                ?? "A success claim is not backed by a passing test."
-        case .red:
-            return reasons.first(where: { $0.rule == .standingTestFailure })?.detail
-                ?? "A test failure was left standing."
-        }
-    }
-}
-
 /// Scans a completed run's transcript (messages + tool calls/results) and classifies it VERIFIED /
 /// UNVERIFIED / RED. Pure, deterministic, and bounded — it never performs I/O, never force-unwraps, and
 /// caps the text it scans — so it is fully unit-testable and can never crash on a hostile transcript.
@@ -344,175 +258,5 @@ public enum RunIntegrityScanner {
         guard text.count > maxClaimScanCharacters else { return text.lowercased() }
         let cut = text.index(text.startIndex, offsetBy: maxClaimScanCharacters)
         return String(text[..<cut]).lowercased()
-    }
-}
-
-/// Persists a `RunIntegrityReport` onto a thread as a `.notice` event so the badge is STABLE across
-/// reloads (the thread is the run, and `JSONThreadStore` round-trips its events). Reuses the notice-event
-/// convention already used for other post-hoc annotations (memory-redaction review) — no schema change.
-public enum RunIntegrityRecord {
-    /// The well-known summary marker that identifies the integrity notice among a thread's events.
-    public static let eventSummary = "run-integrity-report"
-
-    /// The Codable payload stored in the notice event's `payloadJSON`.
-    public struct Payload: Codable, Sendable, Hashable {
-        public var verdict: RunIntegrityVerdict
-        public var summary: String
-
-        public init(verdict: RunIntegrityVerdict, summary: String) {
-            self.verdict = verdict
-            self.summary = summary
-        }
-    }
-
-    /// Builds the notice event for a report. Returns nil only if encoding somehow fails (never throws).
-    public static func event(for report: RunIntegrityReport) -> ThreadEvent? {
-        let payload = Payload(verdict: report.verdict, summary: report.summaryLine)
-        guard let payloadJSON = try? JSONHelpers.encodePretty(payload) else { return nil }
-        return ThreadEvent(kind: .notice, summary: eventSummary, payloadJSON: payloadJSON)
-    }
-
-    /// The most recently recorded integrity verdict on a thread, or nil if none was ever recorded. The
-    /// LAST matching notice wins, so a re-scan after a follow-up turn supersedes an earlier badge.
-    public static func latest(in thread: ChatThread) -> Payload? {
-        for event in thread.events.reversed() where isRecord(event) {
-            if let payloadJSON = event.payloadJSON,
-               let payload = try? JSONHelpers.decode(Payload.self, from: payloadJSON) {
-                return payload
-            }
-        }
-        return nil
-    }
-
-    /// Scans the thread, then records the resulting verdict as a fresh notice event (replacing any prior
-    /// integrity notices so the badge reflects only the latest scan). Returns the report that was
-    /// recorded. Idempotent-ish: repeated calls collapse to a single trailing notice.
-    @discardableResult
-    public static func record(into thread: inout ChatThread) -> RunIntegrityReport {
-        let report = RunIntegrityScanner.scan(thread)
-        thread.events.removeAll(where: isRecord)
-        if let event = event(for: report) {
-            thread.events.append(event)
-            thread.updatedAt = Date()
-        }
-        return report
-    }
-
-    public static func isRecord(_ event: ThreadEvent) -> Bool {
-        event.kind == .notice && event.summary == eventSummary
-    }
-}
-
-/// The data-driven lexicon of what a "test / verify" command looks like. Kept separate so the rule is
-/// inspectable and easy to extend without touching the scanner's control flow. A token matches only on a
-/// word boundary so `pytest` matches but `mypytesting` does not, and `test` alone is intentionally NOT a
-/// token (too many false hits like `test -f`), only test *runners* and `<tool> test` sub-commands.
-public enum TestCommandLexicon {
-    /// Standalone test-runner invocations (word-boundary matched, case-insensitive).
-    public static let runnerTokens: [String] = [
-        "swift test",
-        "xcodebuild test",
-        "xctest",
-        "pytest",
-        "py.test",
-        "unittest",
-        "jest",
-        "vitest",
-        "mocha",
-        "rspec",
-        "phpunit",
-        "gotestsum",
-        "ctest",
-        "tox",
-        "nosetests",
-    ]
-
-    /// `<tool> test` sub-command shapes (e.g. `cargo test`, `npm test`, `go test`, `make test`).
-    public static let subcommandTokens: [String] = [
-        "cargo test",
-        "go test",
-        "npm test",
-        "npm run test",
-        "yarn test",
-        "pnpm test",
-        "bun test",
-        "make test",
-        "make check",
-        "gradle test",
-        "./gradlew test",
-        "mvn test",
-        "dotnet test",
-        "rake test",
-        "bazel test",
-        "ninja test",
-    ]
-
-    public static func looksLikeTestCommand(_ command: String) -> Bool {
-        let lowered = command.lowercased()
-        guard !lowered.isEmpty else { return false }
-        for token in runnerTokens + subcommandTokens where containsToken(token, in: lowered) {
-            return true
-        }
-        return false
-    }
-
-    /// Substring match with cheap word-ish boundaries so `pytest` inside `pytest tests/` matches but a
-    /// token embedded in a longer identifier does not. Bounded, allocation-light.
-    static func containsToken(_ token: String, in haystack: String) -> Bool {
-        var searchRange = haystack.startIndex..<haystack.endIndex
-        while let found = haystack.range(of: token, options: [], range: searchRange) {
-            let beforeOK = found.lowerBound == haystack.startIndex
-                || !isWordScalar(haystack[haystack.index(before: found.lowerBound)])
-            let afterOK = found.upperBound == haystack.endIndex
-                || !isWordScalar(haystack[found.upperBound])
-            if beforeOK && afterOK { return true }
-            if found.upperBound >= haystack.endIndex { break }
-            searchRange = haystack.index(after: found.lowerBound)..<haystack.endIndex
-        }
-        return false
-    }
-
-    static func isWordScalar(_ character: Character) -> Bool {
-        character.isLetter || character.isNumber || character == "_"
-    }
-}
-
-/// The data-driven lexicon of assistant success claims. Anchored so an incidental "passes" in prose does
-/// not trip UNVERIFIED — every phrase reads as a claim ABOUT tests/checks passing. Kept public and
-/// inspectable so the precision stance is auditable and tunable.
-public enum SuccessClaimLexicon {
-    /// Phrases that, appearing verbatim (case-insensitive) in assistant text, assert a checked green.
-    /// Each is deliberately multi-word / anchored to a test/check/build noun to avoid catching the
-    /// English verb "pass" used incidentally (e.g. "the function passes the argument along").
-    public static let claimPhrases: [String] = [
-        "tests pass",
-        "tests passed",
-        "tests are passing",
-        "all tests pass",
-        "all tests passed",
-        "all tests passing",
-        "all tests are passing",
-        "test suite passes",
-        "test suite passed",
-        "tests are green",
-        "all green",
-        "everything passes",
-        "everything is passing",
-        "all checks pass",
-        "all checks passed",
-        "checks pass",
-        "build passes",
-        "build passed",
-        "verified the tests",
-        "tests now pass",
-        "the tests pass",
-    ]
-
-    public static func matchedClaim(in loweredText: String) -> String? {
-        guard !loweredText.isEmpty else { return nil }
-        for phrase in claimPhrases where loweredText.contains(phrase) {
-            return phrase
-        }
-        return nil
     }
 }
